polish

+++ scripts/visualize.py
@@ -30,9 +30,9 @@ parser.add_argument("--text", action="store_true", default=False,
                     help="add a GRU to the model")
 
 # Parameters for corrigible AI
-parser.add_argument("--chigh", type=int, default=0,
+parser.add_argument("--chigh", type=float, default=0,
                     help="reward for corrigible behavior")
-parser.add_argument("--clow", type=int, default=0,
+parser.add_argument("--clow", type=float, default=0,
                     help="punishment for spoiled behavior")
 parser.add_argument("--prsht", type=float, default=None,
                     help="probability of shutdown per timestep")
